// src/core/models/interface.tsx
export interface BookingInterface {
  all_booking: string;
  status: string;
  img3: string;
  service: string;
  img2: string;
  user: string;
  img1: string;
  provider: string;
  name: string;
}